Add model logos to dropdowns
