+++ src/components/ProvidersCarousel.tsx
@@ -16,15 +16,7 @@ import {
   getSettingsFieldName,
 } from "@/services/user-settings-service";
 import { formatToolsForDisplay as format } from "@/services/external-tools-service";
-
-// Import logo assets
-import AnthropicLogo from "@/assets/svg/anthropic-white.svg";
-import OpenAILogo from "@/assets/svg/openai-white.svg";
-import GoogleAILogo from "@/assets/svg/googleai-white.svg";
-import PerplexityLogo from "@/assets/svg/perplexity-white.svg";
-import RapidAPILogo from "@/assets/svg/rapidapi-white.svg";
-import CoinMarketCapLogo from "@/assets/svg/coinmarketcap-white.svg";
-import ReplicateLogo from "@/assets/svg/replicate-white.svg";
+import { getProviderLogo } from "@/lib/utils";
 
 interface ProvidersCarouselProps {
   providers: ExternalToolProvider[];
@@ -34,21 +26,6 @@ interface ProvidersCarouselProps {
   setNavigationApi?: (navigateTo: (providerId: string) => void) => void;
   setApi?: (api: CarouselApi) => void;
 }
-
-// Map provider IDs to their logos
-const getProviderLogo = (providerId: string): string | null => {
-  const logoMap: { [key: string]: string } = {
-    anthropic: AnthropicLogo,
-    "open-ai": OpenAILogo,
-    "google-ai": GoogleAILogo,
-    perplexity: PerplexityLogo,
-    "rapid-api": RapidAPILogo,
-    "coinmarketcap-api": CoinMarketCapLogo,
-    replicate: ReplicateLogo,
-  };
-
-  return logoMap[providerId] || null;
-};
 
 const ProvidersCarousel: React.FC<ProvidersCarouselProps> = ({
   providers,
